+++ src/main/java/ai/deepcode/jbplugin/ui/nodes/ToDoRootNode.java
@@ -7,7 +7,9 @@ import ai.deepcode.jbplugin.ui.ToDoSummary;
 import ai.deepcode.jbplugin.ui.TodoTreeBuilder;
 import com.intellij.ide.util.treeView.AbstractTreeNode;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.ui.Queryable;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -39,8 +41,16 @@ public class ToDoRootNode extends BaseToDoNode{
     return mySummaryNode;
   }
 
+/*
   @Override
   public String getTestPresentation() {
+    return "Root";
+  }
+*/
+
+  @Nullable
+  @Override
+  public String toTestString(@Nullable Queryable.PrintInfo printInfo) {
     return "Root";
   }